Moved BindingModel into the main source code, refactored the code a little, and added a onehot lm
