+++ src/binding_prediction/dataset.py
@@ -1,6 +1,7 @@
 from torch.utils.data import Dataset
 from rdkit.Chem import rdmolops
 from rdkit import Chem
+from .utils import onehot
 import math
 import torch
 import numpy as np
@@ -127,6 +128,7 @@ class DrugProteinDataset(Dataset):
                           bond.GetEndAtomIdx()))
             assert self.bond_dict[str(bond.GetBondType())] != 3
         for atom in mol.GetAtoms():
+            # This could probably be spead up....
             nodes.append(onehot(self.dataset_info['atom_types'].index(atom.GetSymbol()),
                                 len(self.dataset_info['atom_types'])))
 
@@ -230,13 +232,6 @@ class MergeSnE1(object):
 
         sample['features'] = full_features
         return sample
-
-
-def onehot(idx, len):
-    idx = np.array(idx)  # make sure this is an array
-    z = np.array([0 for _ in range(len)])
-    z[idx] = 1
-    return z
 
 
 def collate_fn(batch):
