+++ src/binding_prediction/language_model.py
@@ -1,6 +1,7 @@
 import torch
 import tensorflow as tf
 from binding_prediction.protein import ProteinSequence
+from binding_prediction.utils import onehot
 
 
 class LanguageModel(object):
@@ -26,3 +27,19 @@ class Elmo(LanguageModel):
     def extract(self, x):
         prot = ProteinSequence(x)
         return self.model.predict(prot.onehot).squeeze()
+
+
+class OneHot(LanguageModel):
+    def __init__(self, path):
+        super(OneHot, self).__init__()
+        self.tla_codes = ["A", "R", "N", "D", "B", "C", "E", "Q", "Z", "G",
+                          "H", "I", "L", "K", "M", "F", "P", "S", "T", "W",
+                          "Y", "V"]
+        self.num_words = len(self.tla_codes)
+
+    def extract(self, x):
+        embedding = []
+        for x_i in x:
+            emb_i = [onehot(self.tla_codes.index(w_i), self.num_words) for w_i in x]
+            embedding.append(torch.Tensor(emb_i).float())
+        return embedding
